+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-07-08T21:19:25.625Z",
+    "2026-07-09T07:44:18.154Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {