+++ packages/orq-rc/RELEASES.md
@@ -10649,3 +10649,13 @@ Based on:
 - [typescript v4.12.0-rc.9] packages/orq-rc
 ### Releases
 - [NPM v4.12.0-rc.9] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.9 - packages/orq-rc
+
+## 2026-07-09 07:50:44
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.789.0 (2.916.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.12.0-rc.10] packages/orq-rc
+### Releases
+- [NPM v4.12.0-rc.10] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.10 - packages/orq-rc